pyproject.toml
[build-system]
requires = ["setuptools>=75.8.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "click-and-traverse"
version = "0.1.0"
requires-python = "==3.12.9"
dependencies = [
    "pip",
    "numpy==2.1.3",
    "jax==0.4.38",
    "jaxlie",
    "jax[cuda12]==0.4.38",
    "jaxopt==0.8.5",
    "flax==0.10.4",
    "brax==0.12.3",
    "playground==0.0.4",
    "mujoco==3.3.1",
    "tyro",
    "swanlab>=0.6.12",
    "mediapy",
    "playground",
    "colorlog",
    "tf2onnx",
    "tensorflow==2.19.0",
    "onnxruntime==1.23.2",
    "PyQt5",
    "opencv-python",
    "huggingface-hub>=0.36.0",
    "trimesh>=4.9.0",
    "wandb>=0.22.2",
    "imageio>=2.37.0",
    "evdev>=1.9.2",
    "imageio-ffmpeg>=0.6.0",
    "pytz>=2025.2",
    "black>=25.9.0",
    "ipdb>=0.13.13",
    "absl-py>=2.2.2",
    "osqp>=1.0.5",
    "accelerate==1.11.0",
    "gradio==5.35.0",
    "loop-rate-limiters==1.2.0",
    "ml_collections==1.1.0",
    "pillow==11.3.0",
    "rich",
    "scipy==1.16.3",
    "toml",
    "torch==2.9.0",
    "torchvision==0.24.0",
    "tqdm>=4.67.1",
    "onnxscript>=0.5.6",
    "debugpy>=1.8.17",
    "onnxruntime-gpu==1.23.2",
    "einops>=0.8.1",
    "hydra-core>=1.3.2",
    "scikit-fmm>=2025.6.23",
    "scikit-image>=0.26.0",
]


[tool.setuptools]
packages = ["cat_ppo"]

[tool.ruff]
line-length = 120  # You can match black's default
target-version = "py312"  # Adjust to your Python version

[tool.ruff.lint]
select = ["F"]
ignore = ["F401"]
